Revert "Guaranteed descent"
